WARNING! This function has been deprecated and is no longer supported. Please use the Network_map function. A function to map differentiation statistics.

WARNING! This function has been deprecated and is no longer supported. Please use the Network_map function. A function to map differentiation statistics.

Usage

Dif_Stats_Map(
  dat,
  pops,
  neighbors,
  col,
  breaks = NULL,
  Lat_buffer,
  Long_buffer
)

Arguments

dat

Data frame or character string that supplies the input data. If it is a character string, the file should be a csv. If it is a csv, the 1st row should contain the individual/population names. The columns should also be named in this fashion.

pops

Data frame or character string that supplies the input data. If it is a character string, the file should be a csv. The columns should be named Sample, containing the sample IDs; Population indicating the population assignment of the individual; Long, indicating the longitude of the sample; Lat, indicating the latitude of the sample.

neighbors

Numeric. The number of neighbors to plot connections with.

col

Character vector indicating the colors you wish to use for plotting.

breaks

Numeric. The breaks used to generate the color ramp when plotting. Users should supply 3 values if custom breaks are desired.

Lat_buffer

Numeric. A buffer to customize visualization.

Long_buffer

Numeric. A buffer to customize visualization.

Value

A list containing the map and the matrix used to plot the map.

Author

Keaka Farleigh

Examples

# \donttest{
data(Fst_dat)
Fst <- Fst_dat[[1]]
Loc <- Fst_dat[[2]]
Test <- Dif_Stats_Map(dat = Fst, pops = Loc,
neighbors = 2,
col = c('#fd8d3c','#fc4e2a','#e31a1c','#bd0026','#800026'),Lat_buffer = 1, Long_buffer = 1)# }
#> Registered S3 method overwritten by 'spdep':
#>   method   from
#>   plot.mst ape 
#> Warning: k greater than one-third of the number of data points