A function to estimate heterozygosity and the number of private alleles from a vcf file.

A function to estimate heterozygosity and the number of private alleles from a vcf file.

Usage

Div_stats(VCF, pops, ploidy, write = FALSE, prefix)

Arguments

VCF

Character string indicating the name of the vcf file to be used in analysis.

pops

Character string indicating the name of the population assignment file. This file should have four columns and be in the same order as your vcf file. The first column named Sample indicates the sample name. The second column named Population indicates the population assignment of each individual. The third column named Longitude indicates the longitude of the sample. The fourth column named Latitude indicates the latitude of the sample.

ploidy

Numeric. The ploidy of the data.

write

Boolean. Whether or not to write the output to a file in the current working directory.

prefix

Character string that will be appended to file output.

Value

A list containing the estimated diversity statistics, model output from linear regression of these statistics against latitude, and model plots.

Examples

# \donttest{
data("HornedLizard_Pop")
data("HornedLizard_VCF")
Test <- Div_stats(VCF = HornedLizard_VCF, pops = HornedLizard_Pop,
ploidy = 2, write = FALSE)# }
#> Registered S3 method overwritten by 'GGally':
#>   method from   
#>   +.gg   ggplot2
#> Registered S3 method overwritten by 'genetics':
#>   method      from 
#>   [.haplotype pegas
#> Please note that rgdal will be retired during October 2023,
#> plan transition to sf/stars/terra functions using GDAL and PROJ
#> at your earliest convenience.
#> See https://r-spatial.org/r/2023/05/15/evolution4.html and https://github.com/r-spatial/evolution
#> rgdal: version: 1.6-7, (SVN revision 1203)
#> Geospatial Data Abstraction Library extensions to R successfully loaded
#> Loaded GDAL runtime: GDAL 3.6.2, released 2023/01/02
#> Path to GDAL shared files: C:/Users/Brokenman/AppData/Local/R/win-library/4.3/rgdal/gdal
#>  GDAL does not use iconv for recoding strings.
#> GDAL binary built with GEOS: TRUE 
#> Loaded PROJ runtime: Rel. 9.2.0, March 1st, 2023, [PJ_VERSION: 920]
#> Path to PROJ shared files: C:/Users/Brokenman/AppData/Local/R/win-library/4.3/rgdal/proj
#> PROJ CDN enabled: FALSE
#> Linking to sp version:2.0-0
#> To mute warnings of possible GDAL/OSR exportToProj4() degradation,
#> use options("rgdal_show_exportToProj4_warnings"="none") before loading sp or rgdal.
#> Formatting has finished, moving onto calculations
#> Heterozygosity calculated, moving to private alleles
#> Private Alleles have been calculated, moving onto plotting
#> Calculations have finished, the packages used to perform file formatting and calculations were
#>   vcfR, adegenet, and dartR for formatting, hierfstat to calculate heterozygosity, and poppr to calculate private alleles