A genetic differentiation matrix and locality information for each population. This data was generated by subsetting data of Farleigh et al., 2021.

A symmetric matrix with estimated genetic differentiation (Fst) between 3 populations.

Usage

data(Fst_dat)

Format

A list with two elements:

Fst_dat

Data frame with three rows and three columns

Loc_dat

Data frame containing the locality information for each population

...

Source

Farleigh, K., Vladimirova, S. A., Blair, C., Bracken, J. T., Koochekian, N., Schield, D. R., ... & Jezkova, T. (2021). The effects of climate and demographic history in shaping genomic variation across populations of the Desert Horned Lizard (Phrynosoma platyrhinos). Molecular Ecology, 30(18), 4481-4496.

Examples

data(Fst_dat)
Fst <- Fst_dat[[1]]
Loc <- Fst_dat[[2]]
# \donttest{
 Test <- Network_map(dat = Fst, pops = Loc,
neighbors = 2,col = c('#4575b4', '#91bfdb', '#e0f3f8','#fd8d3c','#fc4e2a'),
statistic = "Fst", Lat_buffer = 1, Long_buffer = 1)# }
#> Registered S3 method overwritten by 'spdep':
#>   method   from
#>   plot.mst ape 
#> Warning: k greater than one-third of the number of data points

Fstat_plot <- Pairwise_heatmap(dat = Fst, statistic = 'FST')