A genetic differentiation matrix and locality information for each population. This data was generated by subsetting data of Farleigh et al., 2021.

A symmetric matrix with estimated genetic differentiation (Fst) between 3 populations.

Usage

data(Fst_dat)

Format

A list with two elements:

Fst_dat

Data frame with three rows and three columns

Loc_dat

Data frame containing the locality information for each population

...

Source

Farleigh, K., Vladimirova, S. A., Blair, C., Bracken, J. T., Koochekian, N., Schield, D. R., ... & Jezkova, T. (2021). The effects of climate and demographic history in shaping genomic variation across populations of the Desert Horned Lizard (Phrynosoma platyrhinos). Molecular Ecology, 30(18), 4481-4496.

Examples

data(Fst_dat)
Fst <- Fst_dat[[1]]
Loc <- Fst_dat[[2]]
# \donttest{
Test <- Dif_Stats_Map(dat = Fst, pops = Loc,
neighbors = 2, countries = c('united states of america', 'mexico'),
col = c('#fd8d3c','#fc4e2a','#e31a1c','#bd0026','#800026'),Lat_buffer = 1, Long_buffer = 1)
#> Warning: k greater than one-third of the number of data points
# }

Fstat_plot <- Pairwise_heatmap(dat = Fst, statistic = 'FST')